requirements.txt
beautifulsoup4==4.13.4
google_api_python_client==2.170.0
pandas==2.2.3
protobuf==6.31.0
pytest==8.3.5
python-dotenv==1.1.0
Requests==2.32.3
SQLAlchemy==2.0.41
